agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.8.0",
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "pytest>=7.0",
    "unidecode>=1.4.0",
    "uvicorn>=0.42.0",
]
